Reorganize files and classes
